# docs/contracts/poundcake-openapi.json
{
  "components": {
    "schemas": {
...
      "DishDetailResponse": {
        "description": "Schema for detailed dish responses (includes recipe).",
        "properties": {
          "actual_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Duration Sec"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "execution_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Ref"
          },
          "execution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Status"
          },
          "expected_duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Duration Sec"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "processing_status": {
            "default": "new",
            "enum": [
              "new",
              "processing",
              "finalizing",
              "complete",
              "failed",
              "abandoned",
              "timeout",
              "canceled"
            ],
            "title": "Processing Status",
            "type": "string"
          },
          "recipe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecipeDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "integer"
          },
          "req_id": {
            "maxLength": 100,
            "title": "Req Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "retry_attempt": {
            "title": "Retry Attempt",
            "type": "integer"
          },
          "run_phase": {
            "default": "firing",
            "enum": [
              "firing",
              "escalation",
              "resolving"
            ],
            "title": "Run Phase",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "req_id",
          "id",
          "recipe_id",
          "retry_attempt",
          "created_at",
          "updated_at"
        ],
        "title": "DishDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PoundCake API",
    "version": "2.0.131"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dishes": {
      "get": {
        "description": "Query Parameters:\n- processing_status: Filter by processing status (new/pending/processing/complete/failed)\n- req_id: Filter by request ID\n- order_id: Filter by order ID (integer)\n- execution_ref: Filter by execution reference\n- limit: Maximum number of results (default: 100, max: 1000)\n- offset: Number of results to skip (default: 0)\n\nReturns 422 Unprocessable Entity if unknown or invalid query parameters are provided.",
        "operationId": "fetch_dishes_api_v1_dishes_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DishDetailResponse"
                  },
                  "title": "Response Fetch Dishes Api V1 Dishes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Dishes",
        "tags": [
          "executor"
        ]
      }
    },
    "/api/v1/dishes/{dish_id}": {
      "patch": {
        "description": "Updates dish status and execution info from chef/timer.",
        "operationId": "update_dish_api_v1_dishes__dish_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Dish",
        "tags": [
          "executor"
        ]
      },
      "put": {
        "description": "Updates dish status and execution info from chef/timer.",
        "operationId": "update_dish_api_v1_dishes__dish_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Dish",
        "tags": [
          "executor"
        ]
      }
    },
    "/api/v1/dishes/{dish_id}/claim": {
      "post": {
        "description": "Atomically claim a new dish for processing.",
        "operationId": "claim_dish_api_v1_dishes__dish_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Dish",
        "tags": [
          "executor"
        ]
      }
    },
    "/api/v1/dishes/{dish_id}/finalize-claim": {
      "post": {
        "description": "Atomically claim a processing dish for finalization.",
        "operationId": "claim_dish_for_finalize_api_v1_dishes__dish_id__finalize_claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Dish For Finalize",
        "tags": [
          "executor"
        ]
      }
    },
    "/api/v1/dishes/{dish_id}/ingredients": {
      "get": {
        "description": "List dish ingredient executions for a dish.",
        "operationId": "list_dish_ingredients_api_v1_dishes__dish_id__ingredients_get",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DishIngredientResponse"
                  },
                  "title": "Response List Dish Ingredients Api V1 Dishes  Dish Id  Ingredients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dish Ingredients",
        "tags": [
          "executor"
        ]
      }
    },
    "/api/v1/dishes/{dish_id}/ingredients/bulk": {
      "post": {
        "description": "Upsert dish ingredient executions for a dish.",
        "operationId": "upsert_dish_ingredients_api_v1_dishes__dish_id__ingredients_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "dish_id",
            "required": true,
            "schema": {
              "title": "Dish Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishIngredientBulkUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpsertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Dish Ingredients",
        "tags": [
          "executor"
        ]
      }
    },
...
  }
}